rework and place things in different files
Makes it easier to understand, plus some comments on what everything
does
